Add files via upload
Okay, so this is just a basic test of an about page.  I've been studying formatting all today, so I'm hoping to produce pictures of "us" -- the anime pictures I used in the powerpoint -- and I'm thinking about adding a footer to the layout page.  I'm working!  It just is something I'm going to be able to dive into more next week.  Expect a better version of about later