Add files via upload
Okay, so this is the first attempt at a layout.  All you'll need to do is replace the templates and static folders to have these work.  Though there will probably be issues in making the tag links in the individual anime page.  I'm not quite happy with the about page (the pictures don't seem to want to center), but it looks alright thus far.